Wrapped package for composer
