+++ README.md
@@ -1,21 +1,27 @@
-laravel-convert-migrations
-==========================
+# Laravel Convert Migrations
 
-This is a custom command to convert your current SQL database schema into a Laravel 4 Migration file.
+This is an artisan command to convert your current SQL database schema into a Laravel 4 Migration file. It'll come really handy when you have started a Laravel project without using migrations, or if you're migrating an old app to Laravel.
 
-Credit to 'bruceoutdoors' Original class = https://gist.github.com/bruceoutdoors/9166186
+## Installation
 
-INSTALLATION
-==========================
-To install this Artisan command, save the above file into your `app/commands` folder.
-Then add `Artisan::add(new ConvertMigrationsCommand);` into your app/start/artisan.php file.
+1. Add the package to your composer.json file and run `composer update`:
 
-USAGE
-==========================
-Now when you run `php artisan` you will see the new `convert:migrations` option.
+```json
+"require": {
+    "adamkearsley/convert-migrations": "dev-master"
+}
+```
 
-You can use this command by typing `php artisan convert:migrations myDatabaseName`.
+2. Add `'Adamkearsley\ConvertMigrations\ConvertMigrationsServiceProvider'` to your `app/config/app.php` file, inside the `providers` array.
 
-If you want to ignore specific tables you can use a comma separated string `--ignore="users, profiles"`.
+## Usage
 
-Example `php artisan convert:migrations myDatabaseName --ignore="table1, table2"`.
+Now it's as easy as running `php artisan convert:migrations myDatabaseName`. Wait a few seconds and, magically, you'll have a new migration in `app/database/migrations`.
+
+**Ignoring Tables**
+
+You can even ignore tables from the migration if you need to. Just use the `ignore` option and separate table names with a comma: `php artisan convert:migrations --ignore="table1, table2"`.
+
+## Credits
+
+Credits go to "bruceoutdoors" [original class](https://gist.github.com/bruceoutdoors/9166186).
