+++ src/Adamkearsley/ConvertMigrations/ConvertMigrationsCommand.php
@@ -1,0 +1,73 @@
+<?php namespace Adamkearsley\ConvertMigrations;
+
+use Illuminate\Console\Command;
+use Symfony\Component\Console\Input\InputOption;
+use Symfony\Component\Console\Input\InputArgument;
+
+class ConvertMigrationsCommand extends Command {
+
+    /**
+     * The console command name.
+     *
+     * @var string
+     */
+    protected $name = 'convert:migrations';
+
+    /**
+     * The console command description.
+     *
+     * @var string
+     */
+    protected $description = 'Converts an existing MySQL database to migrations.';
+
+    /**
+     * Create a new command instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        parent::__construct();
+    }
+
+    /**
+     * Execute the console command.
+     *
+     * @return mixed
+     */
+    public function fire()
+    {
+        $ignoreInput = str_replace(' ', '', $this->option('ignore'));
+        $ignoreInput = explode(',', $ignoreInput);
+        $migrate = new SqlMigrations;
+        $migrate->ignore($ignoreInput);
+        $migrate->convert($this->argument('database'));
+        $migrate->write();
+        $this->info('Migration Created Successfully');
+    }
+
+    /**
+     * Get the console command arguments.
+     *
+     * @return array
+     */
+    protected function getArguments()
+    {
+        return array(
+            array('database', InputArgument::REQUIRED, 'Name of Database to convert'),
+        );
+    }
+
+    /**
+     * Get the console command options.
+     *
+     * @return array
+     */
+    protected function getOptions()
+    {
+        return array(
+            array('ignore', null, InputOption::VALUE_OPTIONAL, 'Tables to ignore', null),
+        );
+    }
+
+}
